plugins/xai/grok_tts/plugins/grok_tts/example/pyproject.toml
[project]
name = "grok-tts-example"
version = "0.0.0"
requires-python = ">=3.10"

dependencies = [
    "python-dotenv>=1.0",
    "vision-agents-plugins-grok-tts",
    "vision-agents-plugins-getstream",
    "vision-agents-plugins-deepgram",
    "vision-agents-plugins-smart-turn",
    "vision-agents-plugins-gemini",
    "vision-agents",
]

[tool.uv.sources]
"vision-agents-plugins-grok-tts" = {path = "..", editable = true}
